+++ zpz.h
@@ -1,18 +1,18 @@
 #ifndef ZPZ_H
 #define ZPZ_H
 
-int zpz_add(int x,int y,int p);
-int zpz_sub(int x,int y,int p);
-int zpz_mul(int x,int y,int p);
-//void euclide_etendu(int* pgcd,int* u,int* v,int a,int b);
-int zpz_inv(int x,int p);
+long int zpz_add(long int x,long int y,long int p);
+long int zpz_sub(long int x,long int y,long int p);
+long int zpz_mul(long int x,long int y,long int p);
+//void euclide_etendu(long int* pgcd,long int* u,long int* v,long int a,long int b);
+long int zpz_inv(long int x,long int p);
 
-int zpzs_add(int x,int y,int p);
-int zpzs_sub(int x,int y,int p);
-int zpzs_mul(int x,int y,int p);
-int zpzs_inv(int x,int p);
+long int zpzs_add(long int x,long int y,long int p);
+long int zpzs_sub(long int x,long int y,long int p);
+long int zpzs_mul(long int x,long int y,long int p);
+long int zpzs_inv(long int x,long int p);
 
-int zpzs_from_zpz(int x,int p);
-int zpz_from_zpzs(int x,int p);
+long int zpzs_from_zpz(long int x,long int p);
+long int zpz_from_zpzs(long int x,long int p);
 
 #endif
